requirements.txt
altair==4.2.2
attrs==22.2.0
beautifulsoup4==4.11.2
blinker==1.5
cachetools==5.3.0
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.1
scikit-learn==1.2.1
scipy==1.10.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4
soynlp==0.0.493
streamlit==1.19.0
streamlit-lottie==0.0.3
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
wordcloud==1.8.2.2
zipp==3.15.0
